Changed the ml model to run only on one core.
